Move getPermissions
