+++ src/utils/api.ts
@@ -1,6 +1,3 @@
-import { Role } from "@/interfaces/Role";
-import { fePermissions } from "@/consts/permissions";
-import { ROLE_HDRUK_SUPERADMIN } from "@/consts/roles";
 import apis from "@/config/apis";
 import config from "@/config/config";
 import { getUserFromToken } from "@/utils/cookies";
@@ -10,63 +7,6 @@ import { Team } from "@/interfaces/Team";
 import { Application } from "@/interfaces/Application";
 import { CohortRequest } from "@/interfaces/CohortRequest";
 import { HttpOptions } from "@/interfaces/Api";
-
-const getPermissions = (
-    userRoles: Role[] = [],
-    teamUserRoles: Role[] | undefined = []
-) => {
-    const enabledUserRoles = userRoles.filter(userRole => userRole.enabled);
-    const enabledTeamUserRoles = teamUserRoles.filter(
-        TeamUserRole => TeamUserRole.enabled
-    );
-
-    const userRoleNames = enabledUserRoles.map(
-        enabledUserRole => enabledUserRole.name
-    );
-    const teamUserRoleNames = enabledTeamUserRoles.map(
-        teamUserRole => teamUserRole.name
-    );
-
-    // combines team user roles and root user roles
-    const allRoles = [...teamUserRoleNames, ...userRoleNames];
-
-    const permissionObj: { [key: string]: boolean } = {};
-
-    // adds all fe.** permissions
-    Object.keys(fePermissions).forEach(key => {
-        if (allRoles.includes(ROLE_HDRUK_SUPERADMIN)) {
-            // gives "hdruk.superadmin" access to all fe.**
-            permissionObj[key] = true;
-        } else {
-            permissionObj[key] = allRoles.some(role =>
-                fePermissions[key as keyof typeof fePermissions]!.includes(role)
-            );
-        }
-    });
-
-    // gets all team user permissions
-    const teamUserPermissions = teamUserRoles
-        .map(enabledUserRole => enabledUserRole.permissions || [])
-        .flat()
-        .map(permission => permission.name);
-
-    // gets all root user permissions
-    const userPermissions = enabledUserRoles
-        .map(enabledUserRole => enabledUserRole.permissions)
-        .flat()
-        .map(permission => permission.name);
-
-    const uniquePermissions = [
-        ...teamUserPermissions,
-        ...userPermissions,
-    ].filter((value, index, array) => array.indexOf(value) === index);
-
-    uniquePermissions.forEach(uniquePermission => {
-        permissionObj[uniquePermission] = true;
-    });
-
-    return permissionObj;
-};
 
 async function get<T>(
     cookieStore: ReadonlyRequestCookies,
@@ -143,11 +83,4 @@ const ThrowPaginationError = (options: HttpOptions | undefined) => {
     }
 };
 
-export {
-    getPermissions,
-    getUser,
-    getTeam,
-    getApplication,
-    getCohort,
-    ThrowPaginationError,
-};
+export { getUser, getTeam, getApplication, getCohort, ThrowPaginationError };
